requirements.txt
asgiref==3.8.1
attrs==24.2.0
django==5.1.1
djangorestframework==3.15.2
dotenv==0.9.9
flake8==7.1.1
iniconfig==2.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
py==1.11.0
pyparsing==3.2.1
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.1.1
pytz==2024.2
sqlparse==0.5.2
toml==0.10.2
